Move loadAllUsersWithProjects from RawlsUserComponent to RawlsBillingProjectComponent
